+++ internal/impl/snowflake/output_snowflake_put_test.go
@@ -194,7 +194,7 @@ snowpipe: '` + tc.snowpipe + `'
 			privateKeyPath:    "resources/ssh_keys/snowflake_rsa_key.p8",
 			stage:             "@test_stage",
 			compression:       "NONE",
-			errConfigContains: "failed to read private key: private key requires a passphrase, but private_key_passphrase was not supplied",
+			errConfigContains: "failed to read private key: private key requires a passphrase, but private_key_pass was not supplied",
 		},
 		{
 			name:           "executes snowflake query without compression",